+++ packages/config/src/android/__tests__/Scheme-test.js
@@ -1,6 +1,15 @@
 import { resolve } from 'path';
-import { getScheme, setScheme } from '../Scheme';
-import { readAndroidManifestAsync } from '../Manifest';
+
+import { formatAndroidManifest, readAndroidManifestAsync } from '../Manifest';
+import {
+  appendScheme,
+  ensureManifestHasValidIntentFilter,
+  getScheme,
+  getSchemesFromManifest,
+  hasScheme,
+  removeScheme,
+  setScheme,
+} from '../Scheme';
 
 const fixturesPath = resolve(__dirname, 'fixtures');
 const sampleManifestPath = resolve(fixturesPath, 'react-native-AndroidManifest.xml');
@@ -28,5 +37,57 @@ describe('scheme', () => {
       return false;
     });
     expect(schemeIntent).toHaveLength(1);
+  });
+});
+
+function removeSingleTaskFromActivities(manifest) {
+  for (let application of manifest.manifest['application']) {
+    for (let activity of application.activity) {
+      if (activity['$']['android:launchMode'] === 'singleTask') {
+        delete activity['$']['android:launchMode'];
+      }
+    }
+  }
+
+  return manifest;
+}
+
+describe('Schemes', () => {
+  it(`ensure manifest has valid intent filter added`, async () => {
+    const manifest = await readAndroidManifestAsync(sampleManifestPath);
+    const manifestHasValidIntentFilter = ensureManifestHasValidIntentFilter(manifest);
+    expect(manifestHasValidIntentFilter).toBe(true);
+  });
+
+  it(`detect if no singleTask Activity exists`, async () => {
+    const manifest = await readAndroidManifestAsync(sampleManifestPath);
+    removeSingleTaskFromActivities(manifest);
+
+    expect(ensureManifestHasValidIntentFilter(manifest)).toBe(false);
+  });
+
+  it(`adds and removes a new scheme`, async () => {
+    const manifest = await readAndroidManifestAsync(sampleManifestPath);
+    ensureManifestHasValidIntentFilter(manifest);
+
+    const modifiedManifest = appendScheme('myapp.test', manifest);
+    const schemes = getSchemesFromManifest(modifiedManifest);
+    expect(schemes).toContain('myapp.test');
+    const removedManifest = removeScheme('myapp.test', manifest);
+    expect(getSchemesFromManifest(removedManifest)).not.toContain('myapp.test');
+  });
+
+  it(`detect when a duplicate might be added`, async () => {
+    let manifest = await readAndroidManifestAsync(sampleManifestPath);
+    ensureManifestHasValidIntentFilter(manifest);
+
+    const modifiedManifest = appendScheme('myapp.test', manifest);
+    expect(hasScheme('myapp.test', modifiedManifest)).toBe(true);
+  });
+
+  it(`detect a non-existent scheme`, async () => {
+    const manifest = await readAndroidManifestAsync(sampleManifestPath);
+
+    expect(hasScheme('myapp.test', manifest)).toBe(false);
   });
 });
